pyproject.toml
[tool.poetry]
name = "weshare"
version = "0.1.0"
description = ""
authors = ["jiwon hwang <realjiwon98@naver.com>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = ">=5.1.5,<6.0.0"
psycopg2 = ">=2.9.10,<3.0.0"
djangorestframework = "^3.15.2"
python-dotenv = ">=1.0.1,<2.0.0"
drf-spectacular = "^0.28.0"
djangorestframework-simplejwt = "^5.4.0"
django-environ = "^0.12.0"
environ = "^1.0"
requests = "^2.32.3"
django-extensions = "^3.2.3"
django-restframework = "^0.0.1"
django-cors-headers = "^4.7.0"
drf-dynamic-fields = "^0.4.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.14.1"
django-stubs = "^5.1.2"
coverage = "^7.6.10"
djangorestframework-stubs = "^3.15.2"
types-jsonschema = "^4.23.0.20241208"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
plugins = ["mypy_django_plugin.main"] # "djangorestframework-stubs"]
python_version = 3.12
strict = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.local"
